+++ black.toml
@@ -10,6 +10,7 @@ exclude = '''
     | \.mypy_cache
     | \.tox
     | \.venv
+    | \.direnv
     | _build
     | buck-out
     | build
