+++ setup.py
@@ -8,16 +8,16 @@ with open("requirements.txt") as f:
     requirements = f.read()
 
 setup(
-    name="pyano",
+    name="pypiano",
     version="0.1.0",
     description="Library to programmatically play piano",
     long_description=readme,
     author="FelixGSE",
     author_email="",
-    url="https://github.com/FelixGSE/pyano",
+    url="https://github.com/FelixGSE/pypiano",
     license="MIT",
-    packages=["pyano"],
-    package_dir={"pyano": "pyano"},
-    package_data={"pyano": ["sound_fonts/*"]},
+    packages=["pypiano"],
+    package_dir={"pypiano": "pypiano"},
+    package_data={"pypiano": ["sound_fonts/*"]},
     install_requires=requirements,
 )
