+++ tox.ini
@@ -13,9 +13,9 @@ description =
 
 commands =
     format: black --config black.toml --check .
-    lint:  flake8 pyano
+    lint:  flake8 pypiano
     unit: pytest tests/
-    mypy: mypy pyano
+    mypy: mypy pypiano
 
 deps =
     format: black
@@ -25,7 +25,7 @@ deps =
 
 [flake8]
 # Exlcude couple directories and files
-exclude = .eggs/*, .tox/*, build/* pyano/__init__.py
+exclude = .eggs/*, .tox/*, build/* pypiano/__init__.py
 statistics = True
 # Yes I know, but 120 is the length for this project. Live with it
 max-line-length=120
